src/main/java/epam/task/gymbootdb/interceptor/RestLoggingInterceptor.java
package epam.task.gymbootdb.interceptor;

import epam.task.gymbootdb.service.LoggingService;

import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;

import lombok.NonNull;
import lombok.RequiredArgsConstructor;

import org.springframework.stereotype.Component;
import org.springframework.web.servlet.HandlerInterceptor;

@Component
@RequiredArgsConstructor
public class RestLoggingInterceptor implements HandlerInterceptor {

    private final LoggingService loggingService;

    @Override
    public boolean preHandle(@NonNull HttpServletRequest request,
                             @NonNull HttpServletResponse response,
                             @NonNull Object handler) {
        String method = request.getMethod();
        String uri = request.getRequestURI();
        String queryString = request.getQueryString() != null ? "?" + request.getQueryString() : "";

        loggingService.logInfoInterceptorPreHandle(method, uri, queryString);

        return true;
    }

    @Override
    public void afterCompletion(@NonNull HttpServletRequest request,
                                @NonNull HttpServletResponse response,
                                @NonNull Object handler,
                                Exception e) {
        int status = response.getStatus();
        String message = (e != null) ? e.getMessage() : "Response OK";

        loggingService.logInfoInterceptorPostHandle(status, message);
    }
}
